+++ tinygl/matrix.cpp
@@ -116,7 +116,7 @@ void glopRotate(GLContext *c,TGLParam *p)
   float angle;
   int dir_code;
 
-  angle = p[1].f * PI / 180.0;
+  angle = (float)(p[1].f * PI / 180.0);
   u[0]=p[2].f;
   u[1]=p[3].f;
   u[2]=p[4].f;
@@ -221,12 +221,12 @@ void glopFrustum(GLContext *c,TGLParam *p)
   float farp=p[6].f;
   float x,y,A,B,C,D;
 
-  x = (2.0*near) / (right-left);
-  y = (2.0*near) / (top-bottom);
+  x = (float)((2.0*near) / (right-left));
+  y = (float)((2.0*near) / (top-bottom));
   A = (right+left) / (right-left);
   B = (top+bottom) / (top-bottom);
   C = -(farp+near) / ( farp-near);
-  D = -(2.0*farp*near) / (farp-near);
+  D = (float)(-(2.0*farp*near) / (farp-near));
 
   r=&m.m[0][0];
   r[0]= x; r[1]=0; r[2]=A; r[3]=0;
